영화 서비스 개선: Result 타입 적용, 테스트 방식 변경 및 DB 타입 불일치 해결

+++ src/movies/entities/movie-theater.entity.ts
@@ -4,7 +4,7 @@ import { Theater } from './theater.entity';
 
 @Entity('movie_theaters')
 export class MovieTheater {
-  @PrimaryGeneratedColumn({ name: 'id', type: 'bigint' })
+  @PrimaryGeneratedColumn({ name: 'id', type: 'integer' })
   id: number;
 
   @ManyToOne(() => Movie, movie => movie.movieTheaters)
